extracting scroll offset calculation duplication
