+++ core/src/main/java/com/novoda/landingstrip/ScrollingPageChangeListener.java
@@ -1,19 +1,20 @@
 package com.novoda.landingstrip;
 
 import android.support.v4.view.ViewPager;
-import android.view.View;
 
 class ScrollingPageChangeListener implements ViewPager.OnPageChangeListener {
 
     private final State state;
     private final TabsContainer tabsContainer;
+    private final ScrollOffsetCalculator scrollOffsetCalculator;
     private final Scrollable scrollable;
 
     private boolean firstTimeAccessed = true;
 
-    ScrollingPageChangeListener(State state, TabsContainer tabsContainer, Scrollable scrollable) {
+    ScrollingPageChangeListener(State state, TabsContainer tabsContainer, ScrollOffsetCalculator scrollOffsetCalculator, Scrollable scrollable) {
         this.state = state;
         this.tabsContainer = tabsContainer;
+        this.scrollOffsetCalculator = scrollOffsetCalculator;
         this.scrollable = scrollable;
     }
 
@@ -49,42 +50,10 @@ class ScrollingPageChangeListener implements ViewPager.OnPageChangeListener {
     }
 
     private void scroll(int position, float positionOffset) {
-        int scrollOffset = getHorizontalScrollOffset(position, positionOffset);
-        float newScrollX = calculateScrollOffset(position, scrollOffset, positionOffset);
-
         state.updatePosition(position);
         state.updatePositionOffset(positionOffset);
 
-        scrollable.scrollTo((int) newScrollX);
-    }
-
-    private int getHorizontalScrollOffset(int position, float swipePositionOffset) {
-        int tabWidth = tabsContainer.getTabAt(position).getWidth();
-        return Math.round(swipePositionOffset * tabWidth);
-    }
-
-    private float calculateScrollOffset(int position, int scrollOffset, float pagerOffset) {
-        View tabForPosition = tabsContainer.getTabAt(position);
-
-        float tabStartX = tabForPosition.getLeft() + scrollOffset;
-
-        int viewMiddleOffset = getTabParentWidth() / 2;
-        float tabCenterOffset = (tabForPosition.getRight() - tabForPosition.getLeft()) * 0.5F;
-
-        float nextTabDelta = getNextTabDelta(position, pagerOffset, tabForPosition);
-
-        return tabStartX - viewMiddleOffset + tabCenterOffset + nextTabDelta;
-    }
-
-    private float getNextTabDelta(int position, float pagerOffset, View tabForPosition) {
-        if (tabsContainer.getTabCount() - 1 >= position + 1) {
-            return (((tabsContainer.getTabAt(position + 1).getWidth()) - tabForPosition.getWidth()) * pagerOffset) * 0.5F;
-        }
-        return 0F;
-    }
-
-    private int getTabParentWidth() {
-        return tabsContainer.getParentWidth();
+        scrollable.scrollTo(scrollOffsetCalculator.calculateScrollOffset(position, positionOffset));
     }
 
     @Override
